refactor: split frame action transitions
